Refactor name
